// src/lexer/componente-lexico.enum.ts
enum ComponenteLexicoEnum {
  L_LLAVE = "L_LLAVE",
  R_LLAVE = "R_LLAVE",
  L_CORCHETE = "L_CORCHETE",
  R_CORCHETE = "R_CORCHETE",
  COMA = "COMA",
  DOS_PUNTOS = "DOS_PUNTOS",
  PR_TRUE = "PR_TRUE",
  PR_FALSE = "PR_FALSE",
  PR_NULL = "PR_NULL",
  LITERAL_CADENA = "LITERAL_CADENA",
  LITERAL_NUM = "LITERAL_NUM",
  ERROR = "ERROR",
  EOF = "EOF",
  EMPTY = "EMPTY"
}


export default ComponenteLexicoEnum